pyproject.toml
[project]
name = "docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.28.1",
    "python-telegram-bot>=22.7",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]
